feat: add cache access to notification widgets

+++ lib/map/models/notification_options.dart
@@ -1,5 +1,5 @@
 class NotificationOptions {
-  final int range;
+  final double range;
   final List<List<double>> coordinates;
 
   NotificationOptions({required this.range, required this.coordinates});
